# api/docs/openapi-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MetivtaEval API",
    "description": "## MetivtaEval - AI Benchmarking Platform\n\nMetivtaEval provides evaluation services for Torah AI systems across DAAT and MTEB modes.\n\n### Features\n\n- **DAAT Scoring**: Deterministic Attribution & Agentic Traceability\n- **MTEB Metrics**: Standard information retrieval metrics\n- **Leaderboard APIs**: Ranked public and user-specific evaluation views\n- **WebSocket Events**: Authenticated progress and leaderboard update streams\n\n### Runtime Notes\n\n- **DAAT Dataset**: Loaded from the configured local JSON dataset in Docker and local dev\n- **LangSmith**: Optional for dataset sync and tracing when credentials are configured\n- **External Evaluators**: Anthropic and Browserless integrations activate when their keys are set\n\n### Authentication\n\nAll protected endpoints accept:\n- **Bearer Token**: JWT in the `Authorization` header\n- **API Key**: `X-API-Key` header",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v2/eval/": {
      "post": {
        "tags": [
          "Evaluation"
        ],
        "summary": "Submit a new evaluation",
        "description": "Submit a new evaluation job.",
        "operationId": "submit_evaluation_api_v2_eval__post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationCreate"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Evaluation"
        ],
        "summary": "List evaluations",
        "description": "List evaluations for current user.",
        "operationId": "list_evaluations_api_v2_eval__get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EvaluationStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EvaluationMode"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mode"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/eval/{evaluation_id}": {
      "get": {
        "tags": [
          "Evaluation"
        ],
        "summary": "Get evaluation status",
        "description": "Get one evaluation status.",
        "operationId": "get_evaluation_api_v2_eval__evaluation_id__get",
        "parameters": [
          {
            "name": "evaluation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Evaluation Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/leaderboard/": {
      "get": {
        "tags": [
          "Leaderboard"
        ],
        "summary": "Get leaderboard",
        "description": "Return ranked leaderboard rows.",
        "operationId": "get_leaderboard_api_v2_leaderboard__get",
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/LeaderboardMode",
              "default": "all"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/leaderboard/{system_id}": {
      "get": {
        "tags": [
          "Leaderboard"
        ],
        "summary": "Get system ranking",
        "description": "Return one leaderboard entry by system/evaluation id.",
        "operationId": "get_system_ranking_api_v2_leaderboard__system_id__get",
        "parameters": [
          {
            "name": "system_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "System Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardEntry"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "EvaluationCreate": {
        "properties": {
          "system_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "System Name"
          },
          "system_version": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "System Version"
          },
          "endpoint_url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Endpoint Url"
          },
          "mode": {
            "$ref": "#/components/schemas/EvaluationMode",
            "default": "daat"
          },
          "dataset_name": {
            "type": "string",
            "title": "Dataset Name",
            "default": "default"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          },
          "async_mode": {
            "type": "boolean",
            "title": "Async Mode",
            "description": "Run evaluation asynchronously",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "system_name",
          "endpoint_url"
        ],
        "title": "EvaluationCreate",
        "description": "Evaluation submission request."
      },
      "EvaluationListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EvaluationResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "EvaluationListResponse",
        "description": "List of evaluations response."
      },
      "EvaluationMode": {
        "type": "string",
        "enum": [
          "daat",
          "mteb"
        ],
        "title": "EvaluationMode",
        "description": "Evaluation mode."
      },
      "EvaluationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "system_name": {
            "type": "string",
            "title": "System Name"
          },
          "system_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System Version"
          },
          "endpoint_url": {
            "type": "string",
            "title": "Endpoint Url"
          },
          "mode": {
            "$ref": "#/components/schemas/EvaluationMode"
          },
          "dataset_name": {
            "type": "string",
            "title": "Dataset Name"
          },
          "status": {
            "$ref": "#/components/schemas/EvaluationStatus"
          },
          "progress": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "type": "object",
        "required": [
          "id",
          "system_name",
          "system_version",
          "endpoint_url",
          "mode",
          "dataset_name",
          "status",
          "progress",
          "created_at",
          "started_at",
          "completed_at"
        ],
        "title": "EvaluationResponse",
        "description": "Evaluation response."
      },
...
      "EvaluationStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "completed",
          "failed",
          "cancelled"
        ],
        "title": "EvaluationStatus",
        "description": "Evaluation status."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LeaderboardEntry": {
        "properties": {
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "system_id": {
            "type": "string",
            "format": "uuid",
            "title": "System Id"
          },
          "system_name": {
            "type": "string",
            "title": "System Name"
          },
          "system_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System Version"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "organization": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "overall_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overall Score"
          },
          "daat_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daat Score"
          },
          "ndcg_10": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ndcg 10"
          },
          "map_100": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Map 100"
          },
          "mrr_10": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mrr 10"
          },
          "dataset_name": {
            "type": "string",
            "title": "Dataset Name"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          }
        },
        "type": "object",
        "required": [
          "rank",
          "system_id",
          "system_name",
          "system_version",
          "author",
          "organization",
          "mode",
          "overall_score",
          "daat_score",
          "ndcg_10",
          "map_100",
          "mrr_10",
          "dataset_name",
          "submitted_at"
        ],
        "title": "LeaderboardEntry",
        "description": "Single leaderboard row."
      },
      "LeaderboardMode": {
        "type": "string",
        "enum": [
          "all",
          "daat",
          "mteb"
        ],
        "title": "LeaderboardMode",
        "description": "Leaderboard mode filter."
      },
      "LeaderboardResponse": {
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/LeaderboardMode"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/LeaderboardEntry"
            },
            "type": "array",
            "title": "Entries"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "last_updated": {
            "type": "string",
            "format": "date-time",
            "title": "Last Updated"
          }
        },
        "type": "object",
        "required": [
          "mode",
          "entries",
          "total",
          "page",
          "page_size",
          "last_updated"
        ],
        "title": "LeaderboardResponse",
        "description": "Leaderboard list payload."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}